feat : 마이페이지 예약 내역 1건 조회 API 마무리

+++ src/main/java/com/zuitopia/petopia/mypage/mapper/MyReservationMapper.java
@@ -8,6 +8,6 @@ import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface MyReservationMapper {
-    ReservationVO getReservationVO(Long userId); // 유저 아이디를 사용해서 예약 1건 가져오기
-    PlaceDTO getReservationPlaceInfo(Long branchId); // 지점 아이디를 사용해서 지점별 반려견 유모차 정보 가져오기
+    ReservationVO getReservationVO(int userId); // 유저 아이디를 사용해서 예약 1건 가져오기
+    PlaceDTO getReservationPlaceInfo(int branchId); // 지점 아이디를 사용해서 지점별 반려견 유모차 정보 가져오기
 }
